Reorganize the scripting API as a separate module
Keep backward compatibility for now, and also add a couple of tests
